Move hibernate managed object boundary to within the DAO itself.
