+++ src/main/java/org/sagebionetworks/bridge/config/SpringConfig.java
@@ -106,7 +106,7 @@ import org.sagebionetworks.bridge.hibernate.BasicPersistenceExceptionConverter;
 import org.sagebionetworks.bridge.json.BridgeObjectMapper;
 import org.sagebionetworks.bridge.models.RequestInfo;
 import org.sagebionetworks.bridge.models.Tag;
-import org.sagebionetworks.bridge.models.assessments.Assessment;
+import org.sagebionetworks.bridge.models.assessments.HibernateAssessment;
 import org.sagebionetworks.bridge.models.files.FileMetadata;
 import org.sagebionetworks.bridge.models.files.FileRevision;
 import org.sagebionetworks.bridge.redis.JedisOps;
@@ -652,7 +652,7 @@ public class SpringConfig {
         metadataSources.addAnnotatedClass(RequestInfo.class);
         metadataSources.addAnnotatedClass(FileMetadata.class);
         metadataSources.addAnnotatedClass(FileRevision.class);
-        metadataSources.addAnnotatedClass(Assessment.class);
+        metadataSources.addAnnotatedClass(HibernateAssessment.class);
         metadataSources.addAnnotatedClass(Tag.class);
         
         return metadataSources.buildMetadata().buildSessionFactory();
