+++ src/test/java/org/sagebionetworks/bridge/hibernate/HibernateAssessmentDaoTest.java
@@ -2,7 +2,8 @@ package org.sagebionetworks.bridge.hibernate;
 
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
-import static org.testng.Assert.assertSame;
+import static org.testng.Assert.assertNotNull;
+import static org.testng.Assert.assertTrue;
 
 import java.util.List;
 import java.util.Map;
@@ -24,6 +25,7 @@ import org.testng.annotations.Test;
 
 import org.sagebionetworks.bridge.models.PagedResourceList;
 import org.sagebionetworks.bridge.models.assessments.Assessment;
+import org.sagebionetworks.bridge.models.assessments.HibernateAssessment;
 
 public class HibernateAssessmentDaoTest extends Mockito {
     
@@ -50,7 +52,7 @@ public class HibernateAssessmentDaoTest extends Mockito {
     private static final String ID_VALUE = "identifier";
     private static final String GUID_VALUE = "guid";
     private static final int REV_VALUE = 3;
-    private static final Assessment ASSESSMENT = new Assessment();
+    private static final HibernateAssessment HIBERNATE_ASSESSMENT = new HibernateAssessment();
     
     @Mock
     HibernateHelper mockHelper;
@@ -62,7 +64,7 @@ public class HibernateAssessmentDaoTest extends Mockito {
     ArgumentCaptor<Map<String,Object>> paramsCaptor;
     
     @Captor
-    ArgumentCaptor<Function<Session, Assessment>> functionCaptor;
+    ArgumentCaptor<Function<Session, HibernateAssessment>> functionCaptor;
     
     @Mock
     Session mockSession;
@@ -76,17 +78,17 @@ public class HibernateAssessmentDaoTest extends Mockito {
         MockitoAnnotations.initMocks(this);
         
         when(mockHelper.executeWithExceptionHandling(any(), any())).then(answer -> {
-            Function<Session,Assessment> func = answer.getArgument(1);
+            Function<Session,HibernateAssessment> func = answer.getArgument(1);
             return func.apply(mockSession);
         });
     }
 
     @Test
     public void getAssessmentsExcludeDeleted() {
-        List<Assessment> list = ImmutableList.of(ASSESSMENT, ASSESSMENT, ASSESSMENT, ASSESSMENT,
-                ASSESSMENT);
+        List<HibernateAssessment> list = ImmutableList.of(HIBERNATE_ASSESSMENT, HIBERNATE_ASSESSMENT, HIBERNATE_ASSESSMENT, HIBERNATE_ASSESSMENT,
+                HIBERNATE_ASSESSMENT);
         when(mockHelper.nativeQueryCount(queryCaptor.capture(), any())).thenReturn(5);
-        when(mockHelper.nativeQueryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(0), eq(20), eq(Assessment.class)))
+        when(mockHelper.nativeQueryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(0), eq(20), eq(HibernateAssessment.class)))
                 .thenReturn(list);
         
         PagedResourceList<Assessment> page = dao.getAssessments(APP_ID_VALUE, 0, 20, null, false);
@@ -102,7 +104,7 @@ public class HibernateAssessmentDaoTest extends Mockito {
     @Test
     public void getAssessmentsIncludeDeleted() {
         when(mockHelper.nativeQueryCount(queryCaptor.capture(), any())).thenReturn(0);
-        when(mockHelper.nativeQueryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(0), eq(20), eq(Assessment.class)))
+        when(mockHelper.nativeQueryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(0), eq(20), eq(HibernateAssessment.class)))
                 .thenReturn(ImmutableList.of());
         
         dao.getAssessments(APP_ID_VALUE, 0, 20, null, true);
@@ -114,8 +116,8 @@ public class HibernateAssessmentDaoTest extends Mockito {
     public void getAssessmentRevisionsIncludeDeleted() {
         when(mockHelper.queryCount(queryCaptor.capture(), paramsCaptor.capture()))
             .thenReturn(100);
-        when(mockHelper.queryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(0), eq(20), eq(Assessment.class)))
-            .thenReturn(ImmutableList.of(ASSESSMENT, ASSESSMENT, ASSESSMENT));
+        when(mockHelper.queryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(0), eq(20), eq(HibernateAssessment.class)))
+            .thenReturn(ImmutableList.of(HIBERNATE_ASSESSMENT, HIBERNATE_ASSESSMENT, HIBERNATE_ASSESSMENT));
         
         PagedResourceList<Assessment> page = dao.getAssessmentRevisions(APP_ID_VALUE, ID_VALUE, 0, 20, true);
         assertEquals(page.getItems().size(), 3);
@@ -130,7 +132,7 @@ public class HibernateAssessmentDaoTest extends Mockito {
 
     @Test
     public void getAssessmentRevisionsExcludeDeleted() {
-        when(mockHelper.queryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(0), eq(20), eq(Assessment.class)))
+        when(mockHelper.queryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(0), eq(20), eq(HibernateAssessment.class)))
             .thenReturn(ImmutableList.of());
         
         dao.getAssessmentRevisions(APP_ID_VALUE, ID_VALUE, 0, 20, false);
@@ -139,11 +141,11 @@ public class HibernateAssessmentDaoTest extends Mockito {
     
     @Test
     public void getAssessmentByGuid() {
-        when(mockHelper.queryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(null), eq(null), eq(Assessment.class)))
-            .thenReturn(ImmutableList.of(ASSESSMENT));
+        when(mockHelper.queryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(null), eq(null), eq(HibernateAssessment.class)))
+            .thenReturn(ImmutableList.of(HIBERNATE_ASSESSMENT));
         
         Optional<Assessment> retValue = dao.getAssessment(APP_ID_VALUE, GUID_VALUE);
-        assertSame(retValue.get(), ASSESSMENT);
+        assertTrue(retValue.isPresent());
         
         assertEquals(queryCaptor.getValue(), HibernateAssessmentDao.GET_BY_GUID);
         
@@ -154,7 +156,7 @@ public class HibernateAssessmentDaoTest extends Mockito {
     
     @Test
     public void getAssessmentByGuidNoEntity() {
-        when(mockHelper.queryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(null), eq(null), eq(Assessment.class)))
+        when(mockHelper.queryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(null), eq(null), eq(HibernateAssessment.class)))
             .thenReturn(ImmutableList.of());
     
         Optional<Assessment> retValue = dao.getAssessment(APP_ID_VALUE, GUID_VALUE);
@@ -164,11 +166,11 @@ public class HibernateAssessmentDaoTest extends Mockito {
 
     @Test
     public void getAssessmentByIdAndRevision() {
-        when(mockHelper.queryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(null), eq(null), eq(Assessment.class)))
-            .thenReturn(ImmutableList.of(ASSESSMENT));
+        when(mockHelper.queryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(null), eq(null), eq(HibernateAssessment.class)))
+            .thenReturn(ImmutableList.of(HIBERNATE_ASSESSMENT));
         
         Optional<Assessment> retValue = dao.getAssessment(APP_ID_VALUE, ID_VALUE, REV_VALUE);
-        assertSame(retValue.get(), ASSESSMENT);
+        assertTrue(retValue.isPresent());
         
         assertEquals(queryCaptor.getValue(), HibernateAssessmentDao.GET_BY_IDENTIFIER);
         
@@ -180,7 +182,7 @@ public class HibernateAssessmentDaoTest extends Mockito {
     
     @Test
     public void getAssessmentByIdAndRevisionNoEntity() {
-        when(mockHelper.queryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(null), eq(null), eq(Assessment.class)))
+        when(mockHelper.queryGet(queryCaptor.capture(), paramsCaptor.capture(), eq(null), eq(null), eq(HibernateAssessment.class)))
             .thenReturn(ImmutableList.of());
         
         Optional<Assessment> retValue = dao.getAssessment(APP_ID_VALUE, ID_VALUE, REV_VALUE);
@@ -189,32 +191,32 @@ public class HibernateAssessmentDaoTest extends Mockito {
 
     @Test
     public void createOrUpdateAssessment() throws Exception {
-        when(mockSession.merge(any())).thenReturn(ASSESSMENT);
+        when(mockSession.merge(any())).thenReturn(HIBERNATE_ASSESSMENT);
         
-        Assessment returnValue = dao.saveAssessment(ASSESSMENT);
-        assertSame(returnValue, ASSESSMENT);
+        Assessment returnValue = dao.saveAssessment(APP_ID_VALUE, new Assessment());
+        assertNotNull(returnValue);
         
-        verify(mockSession).merge(ASSESSMENT);
+        verify(mockSession).merge(any(HibernateAssessment.class));
     }
 
     @Test
     public void deleteAssessment() throws Exception {
-        dao.deleteAssessment(ASSESSMENT);
+        dao.deleteAssessment(APP_ID_VALUE, new Assessment());
         
-        verify(mockSession).remove(ASSESSMENT);
+        verify(mockSession).remove(any());
     }  
     
     @Test
     public void publishAssessment() throws Exception {
         Assessment original = new Assessment();
         Assessment assessmentToPublish = new Assessment();
-        when(mockSession.merge(original)).thenReturn(original);
+        when(mockSession.merge(any())).thenReturn(new HibernateAssessment());
         
-        Assessment retValue = dao.publishAssessment(original, assessmentToPublish);
-        assertEquals(retValue, original);
+        Assessment retValue = dao.publishAssessment(APP_ID_VALUE, original, assessmentToPublish);
+        assertNotNull(retValue);
         
-        verify(mockHelper).executeWithExceptionHandling(eq(original), any());
-        verify(mockSession).saveOrUpdate(assessmentToPublish);
-        verify(mockSession).merge(original);
+        verify(mockHelper).executeWithExceptionHandling(any(HibernateAssessment.class), any());
+        verify(mockSession).saveOrUpdate(any(HibernateAssessment.class));
+        verify(mockSession).merge(any(HibernateAssessment.class));
     }
 }
