+++ src/test/java/org/sagebionetworks/bridge/spring/controllers/SharedAssessmentControllerTest.java
@@ -31,7 +31,6 @@ import org.testng.annotations.Test;
 import org.sagebionetworks.bridge.models.PagedResourceList;
 import org.sagebionetworks.bridge.models.accounts.UserSession;
 import org.sagebionetworks.bridge.models.assessments.Assessment;
-import org.sagebionetworks.bridge.models.assessments.AssessmentDto;
 import org.sagebionetworks.bridge.models.assessments.AssessmentTest;
 import org.sagebionetworks.bridge.services.AssessmentService;
 
@@ -72,7 +71,7 @@ public class SharedAssessmentControllerTest extends Mockito {
         Assessment assessment = AssessmentTest.createAssessment();
         when(mockService.importAssessment(TEST_STUDY_IDENTIFIER, OWNER_ID, GUID)).thenReturn(assessment);
         
-        AssessmentDto retValue = controller.importAssessment(GUID, OWNER_ID);
+        Assessment retValue = controller.importAssessment(GUID, OWNER_ID);
         assertEquals(retValue.getGuid(), assessment.getGuid());
         
         verify(mockService).importAssessment(TEST_STUDY_IDENTIFIER, OWNER_ID, GUID);
@@ -84,7 +83,7 @@ public class SharedAssessmentControllerTest extends Mockito {
         when(mockService.getAssessments(
                 SHARED_STUDY_ID_STRING, 10, 25, STRING_TAGS, true)).thenReturn(page);
         
-        PagedResourceList<AssessmentDto> retValue = controller.getSharedAssessments("10", "25", STRING_TAGS, "true");
+        PagedResourceList<Assessment> retValue = controller.getSharedAssessments("10", "25", STRING_TAGS, "true");
         assertEquals(retValue.getItems().size(), 1);
         assertEquals(retValue.getTotal(), Integer.valueOf(100));
         
@@ -98,7 +97,7 @@ public class SharedAssessmentControllerTest extends Mockito {
         when(mockService.getAssessments(
                 SHARED_STUDY_ID_STRING, 0, API_DEFAULT_PAGE_SIZE, null, false)).thenReturn(page);
         
-        PagedResourceList<AssessmentDto> retValue = controller.getSharedAssessments(null, null, null, "false");
+        PagedResourceList<Assessment> retValue = controller.getSharedAssessments(null, null, null, "false");
         // Just verify this was returned
         assertEquals(retValue.getTotal(), Integer.valueOf(0));
         
@@ -111,7 +110,7 @@ public class SharedAssessmentControllerTest extends Mockito {
         Assessment assessment = AssessmentTest.createAssessment();
         when(mockService.getAssessmentByGuid(SHARED_STUDY_ID_STRING, GUID)).thenReturn(assessment);
         
-        AssessmentDto dto = controller.getSharedAssessmentByGuid(GUID);
+        Assessment dto = controller.getSharedAssessmentByGuid(GUID);
         assertEquals(dto.getGuid(), GUID);
     }
     
@@ -120,7 +119,7 @@ public class SharedAssessmentControllerTest extends Mockito {
         Assessment assessment = AssessmentTest.createAssessment();
         when(mockService.getLatestAssessment(SHARED_STUDY_ID_STRING, IDENTIFIER)).thenReturn(assessment);
         
-        AssessmentDto dto = controller.getLatestSharedAssessment(IDENTIFIER);
+        Assessment dto = controller.getLatestSharedAssessment(IDENTIFIER);
         assertEquals(dto.getIdentifier(), IDENTIFIER);
     }
 
@@ -129,7 +128,7 @@ public class SharedAssessmentControllerTest extends Mockito {
         Assessment assessment = AssessmentTest.createAssessment();
         when(mockService.getAssessmentById(SHARED_STUDY_ID_STRING, IDENTIFIER, 10)).thenReturn(assessment);
         
-        AssessmentDto dto = controller.getSharedAssessmentById(IDENTIFIER, "10");
+        Assessment dto = controller.getSharedAssessmentById(IDENTIFIER, "10");
         assertEquals(dto.getGuid(), GUID);
     }
 
@@ -139,7 +138,7 @@ public class SharedAssessmentControllerTest extends Mockito {
         when(mockService.getAssessmentRevisionsByGuid(
                 SHARED_STUDY_ID_STRING, GUID, 10, 25, true)).thenReturn(page);
         
-        PagedResourceList<AssessmentDto> retValue = controller.getSharedAssessmentRevisionsByGuid(GUID, "10", "25", "true");
+        PagedResourceList<Assessment> retValue = controller.getSharedAssessmentRevisionsByGuid(GUID, "10", "25", "true");
         assertEquals(retValue.getItems().size(), 1);
     }
 
@@ -149,7 +148,7 @@ public class SharedAssessmentControllerTest extends Mockito {
         when(mockService.getAssessmentRevisionsById(
                 SHARED_STUDY_ID_STRING, IDENTIFIER, 0, API_DEFAULT_PAGE_SIZE, false)).thenReturn(page);
         
-        PagedResourceList<AssessmentDto> retValue = controller.getSharedAssessmentRevisionsById(IDENTIFIER, null, null, "false");
+        PagedResourceList<Assessment> retValue = controller.getSharedAssessmentRevisionsById(IDENTIFIER, null, null, "false");
         assertEquals(retValue.getItems().size(), 1);
     }
 
@@ -163,9 +162,9 @@ public class SharedAssessmentControllerTest extends Mockito {
         Assessment assessment = AssessmentTest.createAssessment();
         when(mockService.updateSharedAssessment(eq(TEST_STUDY_IDENTIFIER), any())).thenReturn(assessment);
         
-        mockRequestBody(mockRequest, AssessmentDto.create(assessment));
+        mockRequestBody(mockRequest, assessment);
 
-        AssessmentDto retValue = controller.updateSharedAssessment(GUID);
+        Assessment retValue = controller.updateSharedAssessment(GUID);
         assertEquals(retValue.getIdentifier(), assessment.getIdentifier());
         assertEquals(retValue.getGuid(), assessment.getGuid());
